Move seeder commands to sys & cmp components
