+++ server/app/cli.go
@@ -3,6 +3,8 @@ package app
 import (
 	"context"
 	"fmt"
+	composeCommands "github.com/cortezaproject/corteza/server/compose/commands"
+	"github.com/cortezaproject/corteza/server/pkg/seeder"
 	"sync"
 
 	authCommands "github.com/cortezaproject/corteza/server/auth/commands"
@@ -11,7 +13,6 @@ import (
 	"github.com/cortezaproject/corteza/server/pkg/api/server"
 	"github.com/cortezaproject/corteza/server/pkg/cli"
 	"github.com/cortezaproject/corteza/server/pkg/options"
-	fakerCommands "github.com/cortezaproject/corteza/server/pkg/seeder/commands"
 	"github.com/cortezaproject/corteza/server/store"
 	systemCommands "github.com/cortezaproject/corteza/server/system/commands"
 	"go.uber.org/zap"
@@ -132,7 +133,10 @@ func (app *CortezaApp) InitCLI() {
 		federationCommands.Sync(ctx, app),
 		cli.EnvCommand(),
 		cli.VersionCommand(),
-		fakerCommands.Seeder(ctx, app),
+		seeder.BaseCommand(
+			systemCommands.SeedUsers(ctx, app),
+			composeCommands.SeedRecords(ctx, app),
+		),
 	)
 
 }
